+++ src/Either.ts
@@ -1036,9 +1036,6 @@ export function getWitherable<E>(M: Monoid<E>): Witherable2C<URI, E> {
     partitionMap: F_.partitionMap,
     traverse: traverse_,
     sequence,
-    reduce: reduce_,
-    foldMap: foldMap_,
-    reduceRight: reduceRight_,
     wither,
     wilt
   }
@@ -1145,9 +1142,6 @@ export const Foldable: Foldable2<URI> = {
 export const Traversable: Traversable2<URI> = {
   URI,
   map: map_,
-  reduce: reduce_,
-  foldMap: foldMap_,
-  reduceRight: reduceRight_,
   traverse: traverse_,
   sequence
 }