+++ src/ReadonlyArray.ts
@@ -2102,9 +2102,6 @@ export const FoldableWithIndex: FoldableWithIndex1<URI, number> = {
 export const Traversable: Traversable1<URI> = {
   URI,
   map: map_,
-  reduce: reduce_,
-  foldMap: foldMap_,
-  reduceRight: reduceRight_,
   traverse: traverse_,
   sequence
 }
@@ -2141,9 +2138,6 @@ export const Witherable: Witherable1<URI> = {
   filterMap: filterMap_,
   partition: partition_,
   partitionMap: partitionMap_,
-  reduce: reduce_,
-  foldMap: foldMap_,
-  reduceRight: reduceRight_,
   traverse: traverse_,
   sequence,
   wither: wither_,