+++ CHANGELOG.md
@@ -111,6 +111,8 @@ high state of flux, you're at risk of it changing without notice.
     - make `leftOrBoth` and `rightOrBoth` lazy (@gcanti)
   - `Traced`
     - make `tracks` curried (@gcanti)
+  - `Traversable`
+    - drop `Foldable` dependency (@gcanti)
   - `Tree`
     - make `elem` curried (@gcanti)